+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 * SPOTenantCdnPolicy
   * If properties in the tenant are empty then export them as empty arrays
     instead of null strings, missed while fixing #4658
+* SPOTenantSettings
+  * Remove property UserVoiceForFeedbackEnabled when setting the resource since
+    it has been deprecated
 * M365DSCUtil
   * Fixed an issue in `Assert-M365DSCBlueprint` where the clone and export
     of a blueprint with a GUID could lead to configuration name starting